Refactor tcms.core.migrations.0001_django_comments__object_pk
this migration is modifying a 3rd party model and was overriding
app_label to django_comments in order to gain access to the model!

However due to how Django records executed migrations it was
recorded as django_comments|0001... instead of core|0001...! Thus
the migration was seen as still pending and could lead to problems
when applying further migrations.

Signed-off-by: Mr. Senko <atodorov@mrsenko.com>